+++ src/main/java/turing/turing/domain/comment/CommentRepository.java
@@ -1,0 +1,10 @@
+package turing.turing.domain.comment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import turing.turing.domain.question.Question;
+
+import java.util.List;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    List<Comment> findAllByQuestion(Question question);
+}
